+++ cch/src/main/java/wft/hahn/neo4j/cch/search/ExtendedSearchPath.java
@@ -139,7 +139,7 @@ public class ExtendedSearchPath implements SearchPath {
     }
 
     @Override
-    public float weight() {
+    public int weight() {
         return start.weight() + lastArc.weight;
     }
