+++ cch/src/test/java/wtf/hahn/neo4j/cch/storage/BufferManagerTest.java
@@ -1,28 +1,17 @@
 package wtf.hahn.neo4j.cch.storage;
 
-import static java.lang.Float.parseFloat;
-import static java.lang.Integer.parseInt;
-import static org.assertj.core.api.Assertions.assertThat;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.io.TempDir;
+import wft.hahn.neo4j.cch.storage.*;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.List;
-import java.util.Map;
-import java.util.Random;
+import java.util.*;
 import java.util.stream.Collectors;
 
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-import wft.hahn.neo4j.cch.storage.ArcWriter;
-import wft.hahn.neo4j.cch.storage.DiskArc;
-import wft.hahn.neo4j.cch.storage.FifoBuffer;
-import wft.hahn.neo4j.cch.storage.Mode;
-import wft.hahn.neo4j.cch.storage.PositionWriter;
+import static java.lang.Integer.parseInt;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class BufferManagerTest {
 
@@ -34,7 +23,7 @@ public class BufferManagerTest {
         Map<Integer, List<DiskArc>> arcGroup = Files.lines(Paths.get("src", "test", "resources", "rome99.csv"))
                 .map(line -> Arrays.stream(line.split(",")))
                 .map(stream -> stream.map(String::trim).toArray(String[]::new))
-                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseFloat(line[2])))
+                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseInt(line[2])))
                 .filter(diskArc -> diskArc.start() < diskArc.end())
                 .collect(Collectors.groupingBy(DiskArc::start));
         int sum = arcGroup.values().stream().mapToInt(Collection::size).sum();
@@ -56,7 +45,7 @@ public class BufferManagerTest {
         Map<Integer, List<DiskArc>> arcGroup = Files.lines(Paths.get("src", "test", "resources", "rome99.csv"))
                 .map(line -> Arrays.stream(line.split(",")))
                 .map(stream -> stream.map(String::trim).toArray(String[]::new))
-                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseFloat(line[2])))
+                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseInt(line[2])))
                 .filter(diskArc -> diskArc.start() < diskArc.end())
                 .distinct()
                 .collect(Collectors.groupingBy(DiskArc::start));
@@ -84,7 +73,7 @@ public class BufferManagerTest {
         Map<Integer, List<DiskArc>> arcGroup = Files.lines(Paths.get("src", "test", "resources", "rome99.csv"))
                 .map(line -> Arrays.stream(line.split(",")))
                 .map(stream -> stream.map(String::trim).toArray(String[]::new))
-                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseFloat(line[2])))
+                .map(line -> new DiskArc(parseInt(line[0]), parseInt(line[1]), -1, parseInt(line[2])))
                 .filter(diskArc -> diskArc.start() > diskArc.end())
                 .distinct()
                 .collect(Collectors.groupingBy(DiskArc::end));
